Split up utils.js and its tests
Split up the functions in utils.js and and utils.test.js into files for
each function.
