+++ test/lib/getAllFeatureFlags.test.js
@@ -1,0 +1,5 @@
+describe("lib/getAllFeatureFlags", () => {
+    it("exports functions", () => {
+        expect("getAllFeatureFlags").toBeDefined();
+      });
+});
